skip OIDC authentication flow if 'authorization' is part of the additional headers
