+++ weaviate/util.py
@@ -6,8 +6,7 @@ import json
 import base64
 import warnings
 import uuid as uuid_lib
-from typing import Union, Sequence, Tuple, Any, Optional
-from numbers import Real
+from typing import Union, Sequence, Any, Optional
 from io import BufferedReader
 import validators
 import requests
@@ -436,45 +435,6 @@ def _compare_properties(sub_set: list, set_: list) -> bool:
         if not found:
             return False
     return True
-
-
-def _get_valid_timeout_config(timeout_config: Union[Tuple[Real, Real], Real, None]):
-    """
-    Validate and return TimeOut configuration.
-
-    Parameters
-    ----------
-    timeout_config : tuple(Real, Real) or Real or None, optional
-            Set the timeout configuration for all requests to the Weaviate server. It can be a
-            real number or, a tuple of two real numbers: (connect timeout, read timeout).
-            If only one real number is passed then both connect and read timeout will be set to
-            that value.
-
-    Raises
-    ------
-    TypeError
-        If arguments are of a wrong data type.
-    ValueError
-        If 'timeout_config' is not a tuple of 2.
-    ValueError
-        If 'timeout_config' is/contains negative number/s.
-    """
-
-    if isinstance(timeout_config, Real) and not isinstance(timeout_config, bool):
-        if timeout_config <= 0.0:
-            raise ValueError("'timeout_config' cannot be non-positive number/s!")
-        return (timeout_config, timeout_config)
-
-    if not isinstance(timeout_config, tuple):
-        raise TypeError("'timeout_config' should be a (or tuple of) positive real number/s!")
-    if len(timeout_config) != 2:
-        raise ValueError("'timeout_config' must be of length 2!")
-    if not (isinstance(timeout_config[0], Real) and isinstance(timeout_config[1], Real)) or \
-            (isinstance(timeout_config[0], bool) and isinstance(timeout_config[1], bool)):
-        raise TypeError("'timeout_config' must be tuple of real numbers")
-    if timeout_config[0] <= 0.0 or timeout_config[1] <= 0.0:
-        raise ValueError("'timeout_config' cannot be non-positive number/s!")
-    return timeout_config
 
 
 def generate_uuid5(identifier: Any, namespace: Any = "") -> str:
